29595: [Build] Adjusted build to take the new zip file name of the TICS IntelliJ plugin into account.
Previously, the plugin would be packaged as TICSIntelliJ.zip, and we would rename this to TICSIntelliJ-[version].zip. Now, this TICSIntelliJ-[version].zip is produced directly by the Gradle build, so we can simply copy it (for publishing purposes). In order for the TICS Installer to be able to still find TICSIntelliJ.zip as well, we provide a separate copy for that purpose.

Reviewed by @zon.