#31903: fixed issue where violations are not reported + error
For some reason with Joeps machine the xml was returned with backslashes, where forward slashes were expected by the IntelliJ virtual file. Normalizing both the paths now resolves that issue. Also fixed an issue where parsing an OpenFileDescriptor to VirtualFile was not possible by removing a not needed datakey in FileTree.