#32999: Changed analysis behaviour
Analysis used to run according to context, which meant that having the explorer in focus meant the selected file would be analyzed instead of the file opened in the editor. The behaviour has now changed to analyzing the file opened in the editor unless triggered by right clicking the file in the explorer.